Fix issue in 'pdc' that was showing empty orphan nodes ##print

+++ libr/include/r_vector.h
@@ -89,7 +89,7 @@ static inline void *r_vector_index_ptr(RVector *vec, size_t index) {
 }
 
 static inline void *r_vector_at(RVector *vec, int index) {
-	if (vec && index >= 0 && index < vec->capacity) {
+	if (vec && index >= 0 && (size_t)index < vec->capacity) {
 		return (char *)vec->a + (vec->elem_size * index);
 	}
 	return NULL;
